Conditional profiling for top-level definitions
Summary:
While debugging performance problems in large files, I find myself spending a lot of time bisecting the file to see where the regression is. This diff adds the `--profile-toplevel-definitions` option to `hhstc` in order to break down the time it takes to typecheck by top-level entities.

Because class definitions are too coarse, we don't report on them.

Reviewed By: CatherineGasnier

Differential Revision: D34121506

fbshipit-source-id: 4aa830b1cbd4671cec92ebc37056c36b485dd130