Refactor Label iteration
Summary:
A number of different bytecode instructions have one or more Labels to represent branch targets.
The hackc emitter needs to visit all Labels for various reasons; the `Instruct ADT` has them scattered
around in different places making it verbose and error prone to visit them correctly.

This diff introduces an idiom we used in the hacknative IR for operands and branch targets
that is fast and encapsulated. We provide accessors that return slices, either `&[Label]` or
`&mut [Label]`, which only have to do the careful pattern match in one place.
Now client code can work with slices (fast), and pattern matching is simplified.

Iterating over a slice is easy (a for loop or iterator chain). If the visitor is a lambda, it is only
called or instantiated in one place, which simplifies code. Transposing
SSwitch's `Vec<Pair(Str,Label)>` to { Vec<Str>, Vec<Label> }, avoids the need for a
transient Vec<Label> in label_rewriter.

Clarify naming a little bit: a Label operand to a conditional branch instruction is a *target*,
not to be confused with a `ILabel(Label)` instruction which marks the location conditional
branches jump to.

Reviewed By: aorenste

Differential Revision: D34112173

fbshipit-source-id: 8ae82e96a855659f083fbcbffcf1b152b9f46ea0