refactor symbol_builder_types
Summary:
`Symbol_builder_types` defines several independent public types, while functions operating on these types are spread in the other modules.

The motivation for this refactoring is to regroup types with their related functions.  This reduces the number of import and dependencies between modules and allow to make some types abstract.

This diff

1. make the `progress` type in `symbol_builder_types` abstract by moving it along side the functions that operate on it in `symbol_add_fact`
2. move `file_info` to a separate module
3. Rename `symbol_builder_type` to `symbol_predicate`
4. update the add_fact unit test so it uses only the exposed API

Reviewed By: simonmar

Differential Revision: D34125697

fbshipit-source-id: 93dd77b7417233a2b71ae6d570535b8d99199a39