fill in (some) missing substitution stuff
Summary: add `tparams` to folded classes. this is enough to enable us to complete `inherit::make_substitution` and fill out a real implementation for `substituion::Subst::new`. that's not the end of the story, we now to implement `substitution::Subst::instantiate` but that's a whole 'nother module and so for a future diff.

Reviewed By: enetsee

Differential Revision: D34154122

fbshipit-source-id: f34aecc8f248814926f70c7298c8eba0749c9a6b