Always inline coeffects backdoor and the lambda it calls
Summary:
In order to make coeffect migration possible without having to change the entire codebase all at one, we have added backdoor functions. While individual parts of the code are being migrated, it is common to take code that looks like
```
function foo() [defaults] {
  call_defaults();
  call_policied();
}
```
and convert the `foo` to `policied` by
```
function foo() [policied] {
  // Blocked on migration
  HH\Coeffects\backdoor(()[defaults] ==> call_defaults());
  call_policied();
}
```
However, doing this pattern more often than not result in breaking inlining and what's worse is in hot code, results in regressions.

This diff aims to convert this code pattern into a zero cost abstraction by always inlining the backdoor function and the lambda it creates. This is always safe to do since that's exactly what the original code used to look like. Note that the newly created closure is also 1:1 caller::callee so it is equally safe to inline.

Differential Revision: D34049715

fbshipit-source-id: e3c03c6e4bf29fe3bffe237d751a40680b5241bd