Introduce `Intersect` typing errors
Summary:
The current side effecting implementation of errors makes use of FIXME error suppression when typing intersections. If one or more elements of an intersection can be typed without generating errors, either because there are none or they are suppressed, the whole intersection will generate no errors.

As we move towards pure errors, we want to delay any error suppression to the reporting stage. This is at odds with the current mechanism since if we no longer suppress and error and it was the only component which would have type check with that suppression, we raise the errors for all of the components of the intersection.

To work around this, this diff introduces a new `Intersect` constructor for `Typing_error.t` and updates the evaluator to generate a `Eval_result.t` of `User_error.t` results which retains the information about which errors came from an intersection.

Once we move to pure errors, the intention is to generate this intersection errors in typing and consider a list of errors to be suppressed by a FIXME if one element if suppressed.

The main changes here are:
1) evaluating to to a `User_error.t Eval_result.t`; and
2) suppressing intersections of errors given some function which tells us whether an individual error is suppressed

Reviewed By: vsiles

Differential Revision: D34175562

fbshipit-source-id: 87dbe28c41f19d13e37b19338673b7c6ccad5aea