Introduce `Multiple` and `Union` constructors for `Typing_error.t`
Summary:
This diff introduces support for multiple errors on an expression at a single type (`Multiple`) and errors for an expression across a union of types (`Union`).

This is necessary since we want to be able to represent intersections of errors (i.e. errors for an expression at multiple types) which must allow for many errors for each component

Reviewed By: vsiles

Differential Revision: D34196509

fbshipit-source-id: 464b7eb06f422200b8bd894c2f8f12b71b5ce722