Remove FP chain from BeginInlining
Summary:
The FP chain in begin inlining is unecessary.  Things use it to compute offsets
that could be burnt into the extra data of the begin inlining.

This diff does still makes use of the chain for non offset purposes (by walking
the copy stored in the bytecode marker).  In pacticular the fp chain is useful
to be able to produce `(src_key, inlining_depth)` pairs for the whole call
chain from a single IR instruction.  We could do this by cramming all that info
in the extra_data as well, but walking the chain is more convenient.

Reviewed By: ricklavoie

Differential Revision: D34186786

fbshipit-source-id: 33eb0ce84c9aa3cf7139467ef38f3b75badc176c