Don't allow XHP names for module decls & enum classes
Summary: `require_class_name`, a parser utility, allows XHP class names, which might not be immediately obvious. Instead let's rename the utility to indicate it might use a class name. On top of this, we change enum classes and module declarations to require *names* rather than class names. This is what enums do, for example.

Reviewed By: vsiles

Differential Revision: D34183970

fbshipit-source-id: 793bd2bfc98f7541cb1c5691080c58bcd0f81411