defend against buck2 "Text file is busy"
Summary:
See https://fb.workplace.com/groups/buck2users/posts/3078672432389117/

Buck2 is currently vulnerable to "Text file is busy" which usually means that a process still has an FD with a write handle open on the executable, at the time we try to invoke it. T107518211 has plans to defend against it.

In the meantime we're getting several test failures in buck2 because of it.

What other people do is they catch "Text file is busy" (manifest as exit code 126, "command is found but not executable"), and in response they wait and then retry. But I don't know. Implementing retry logic feels like a pain, and every retry loop indeed also needs the ability to give up after some number of retries. Therefore: what I've done in this diff is only catch the problem in the cases where it has manifest so far (rather than a cross-cutting solution), and I've hard-coded in whatever number of retries involved the least code disruption (1 and 0, in the two cases I'm looking at).

Reviewed By: shayne-fletcher

Differential Revision: D34190504

fbshipit-source-id: b9ebd95f8c7872c2f81c0885c66aca29f6002c71