Remove self-inclusion of "typed-value.h" header
Summary: In D29619639 (https://github.com/facebook/hhvm/commit/40584aca54e13db34a8197972c102d74ce881846), "typed-value.h" was altered to include itself. While this is safe because of the `once` pragma guard, this brings up unnecessary warnings in the "Problems" section of VSCode as well as inline in the VSCode editor.

Reviewed By: ricklavoie

Differential Revision: D34226969

fbshipit-source-id: d0ba6f7de2a36d8ab812227ed8a428cd744d9287