Remove ft_arity and fun_arity type
Summary:
Instead of having a separate "arity" for variadic function types, simply use the last parameter and the recently-added flag on the function type flags.

As can be seen from the tests, this diff actually fixes a soundness bug with variadic lambdas, whereby `(...$x) ==> $x[0]->foo()` would be accepted because `$x` would be given a `vec<_>` type. It is now correctly given a type `vec<#1>` for some type variable `#1`.

Some of the clunky List.last and List.drop_last will be cleaned up in a subsequent diff.

Reviewed By: Wilfred

Differential Revision: D34043718

fbshipit-source-id: af9777b134ad109ea008c0c1fe5487bde910427f