Remove TODO comments
Summary:
What I left as a TODO was going to be impossible. I was planning to get rid of the `malloc/free`s but that's quite impossible because some of the `malloc`s are done by the OCaml runtime.

To keep the interface simple (e.g. unconditional `free` in the destructor), I didn't convert any of the mallocs.

Reviewed By: CatherineGasnier

Differential Revision: D34086784

fbshipit-source-id: 914edc503b9e680a0e23da3021b500f941290ca4