Fix shmffi_move and remove write_map
Summary:
Previously the `shmffi_move` implementation was incorrect. It contained a race condition, where a pointer to some data was alive, while a lock was not held.

With the new `inspect_and_remove` functionality, this is no longer possible.

This diff reimplements `shmffi_move` to use that functionality. This involves copying the bytes to a local buffer in between removal and re-insertion.

Reviewed By: CatherineGasnier

Differential Revision: D34086492

fbshipit-source-id: 15383c1fe25e6c7d45c5a7a8f43f7d091b162594