reset type variable identifiers for each top-level definition
Summary:
So far, type variable identifier were created with `Ident.tmp`. `Ident` which is a global, worker-scoped identifier provider. This would cause type variable identifier to be non-deterministic and caused issues.

In this diff, I added a type-var specific identifier provider in Typing_inference_env.

Initially I wanted to make it a field in the inference environment, but was afraid of tricky environment wiring along the control flow, which could have the disastrous effect of creating two different type variables with the same identifier.

Instead, I keep the identifier provider global and reinitialise it at each top-level definition.

Ideally, the new `identifier` type should be abstract, but global inference makes it difficult to do.

Reviewed By: madgen

Differential Revision: D34211990

fbshipit-source-id: 076e038319ccd72e6623e9f0cb122fe2e5e6a0c1