Introduce SharedProfile for non-thread-local profiling
Summary:
SharedProfile is a near-replacement for TargetProfile that does not use RDS to store its data. Instead, it stores the profile in shared memory and exposes a closure-based update to operate on it. The goal is to save on profiling memory, so that we can create more complex profiles, starting with a recursive DecRefProfile. Note that unlike TargetProfile, we call the constructor of profiles allocated by SharedProfile, so they can include data structures like hash tables.

Right now, SharedProfile is implemented with a simple mutex on each profile entry. We can make this implementation more complex in the future, such as by tracking lock contention and allocating more copies as needed. The abstraction will let us introduce these other implementations without changing any usage sites.

Reviewed By: ricklavoie

Differential Revision: D33174281

fbshipit-source-id: cb0f9e48bebb074f18781e89fd9a2515cdf0ae45