Use iface vtable slots for C++ classof checks
Summary:
Here's another optimization for "is class A a subclass of class B" checks. In repo mode, we assign interfaces B a "vtable index" such that any two interfaces implemented by the same class have different vtable indices. (Doing so is a standard graph coloring problem that can be solved greedily.)

To implement classof when A is a non-interface and B is an interface, we can use this vtable index. Again, JIT-ed code already uses this approach in some cases, but now we get the same optimization in C++ as well.

This change knocks another 25% (relative to the base of this stack) off the runtime of classof. The whole stack saves a bit more than 50%.

Reviewed By: ricklavoie

Differential Revision: D32896917

fbshipit-source-id: fca0d0e545ac5a14114f6ea728d3e0565c2dfc77