Use InstanceBits for C++ classof checks
Summary:
HHVM has a PGO-based optimization for "is class A a subclass of class B" checks. We use profiling to identify the top 128 most commonly checked parent classes (class B, in our example), then initialize a bitset in every classes tracking membership in these parents' subtrees.

Today, we only apply this optimization in the JIT. We can apply the same optimization to the C++ helper, as well. The C++ helper is used for interpreted cases, and, more importantly, in the middle of deep TypeStructure type checks.

To keep the C++ code fast, we want to store a class's bitset index (if any) in the class itself. There's enough space in the class to store it there. There's also a small race condition we have to take care of before storing it: we must initialize all class's bitsets before initializing any class's bitset index, because as soon as we set the bitset index, the classof check will start using it.

The effect of this change is entirely restricted to a C++ runtime helper, so we can see it in perf profiles - it makes classof about 25% faster.

Reviewed By: ricklavoie

Differential Revision: D32895624

fbshipit-source-id: f482d2c1575c71de76b1fb2dd9ffc596fef94699