Use a circular buffer for the stack
Summary:
In some runs I was seeing `taint::iopPreamble` now take 20% or more of
the CPU time. Most of this looked to be in `std::deque` operations, and
allocations there.

I dug around the HHVM stack code and we have a fixed size stack (IIRC, 16K) -
so piggy back off of that and write a bog standard circular queue with a fixed
size. This takes allocations off the critical path.

Reviewed By: dkgi

Differential Revision: D34247804

fbshipit-source-id: 0961d7f45f2e77b679b6584799c821fd64f7fcbc