Share more enums in hhbc-ast
Summary:
Shared enums compile to a C++ enum class with a sufficiently sized integral base type decided by CXX. Because it is not UB in C++ for an enum class to hold a value different from all of the listed variants, the rust representation of this is not a Rust enum. Rather its a set of constant struct values that represent these values. Because of this, we have to introduce wildcard fallbacks for the switches that use them.
```
// generated header

struct PlayingCard final {
  Suit suit;
  uint8_t value;
};

enum class Suit : uint8_t {
  Clubs = 0,
  Diamonds = 1,
  Hearts = 2,
  Spades = 3,
};
```
```
// Rust representation

#[derive(Copy, Clone, PartialEq, Eq)]
#[repr(transparent)]
pub struct Suit {
    pub repr: u8,
}
#[allow(non_upper_case_globals)]
impl Suit {
    pub const Clubs: Self = Suit { repr: 0 };
    pub const Diamonds: Self = Suit { repr: 1 };
    pub const Hearts: Self = Suit { repr: 2 };
    pub const Spades: Self = Suit { repr: 3 };
}
```
>Pattern matching with match still works but will require you to write wildcard arms to handle the situation of an enum value that is not one of the listed variants.
```
fn main() {
    let suit: Suit = /*...*/;
    match suit {
        Suit::Clubs => ...,
        Suit::Diamonds => ...,
        Suit::Hearts => ...,
        Suit::Spades => ...,
        _ => ...,  // fallback arm
    }
}
```
https://fburl.com/1vhraw73

I think having these types match and avoiding maps between identical types makes this still a good idea,

Reviewed By: shayne-fletcher

Differential Revision: D34258986

fbshipit-source-id: f89f7801b79cb087d60f3ce59dc658d26683b5f6