Undeclare HH\class_get_class_name
Summary: This is too easily confused with `get_class`, which operates on objects. The loose typing (because we don't have a class pointer type) combined with autocomplete makes it all the more likely that people will unwittingly produce fatals because `HH\class_get_class_name` is expecting a class pointer, not an object.

Reviewed By: arnabde03

Differential Revision: D34336907

fbshipit-source-id: 75e3d508fc0f4a780fde40fcbcec596423a8ef67