Remove InstrSeq in HhasProperty (1/3)
Summary:
In the process of compiling class properties, we must sometimes generate
initializer code fragments. These fragments were stashed in HhasProperty
and later merged into an 86pinit or 86sinit method. Since HhasProperty is meant
to be the final result and is part of the cbindgen-based C++ api, passing the instruction
fragments along separately is the way to go.

This deletes the first of three places InstrSeq shows up in HackCUnit.

Reviewed By: kavoor

Differential Revision: D34222106

fbshipit-source-id: 6fe731f260b4cf9a82571a193301df9a1dc0bb1b