Remove InstrSeq from HhasConstant (2/3)
Summary:
Similar to HhasProperty, a HhasConstant sometimes needs initialization instructions.
These always make their way into a generated method, so when compilation is
finished there is no reason for HhasConstant to have an InstrSeq field; we can pass
it along separately during compilation.

This deletes the second of three places where InstrSeq shows up in HackCUnit.

While we're here, move from_ast() from hhas_constant.rs, which is an ADT module,
to emit_constant.rs which is where the other emitter code is.

Reviewed By: aorenste

Differential Revision: D34342901

fbshipit-source-id: 9716917ab34bef4380845655b5b808cd49f9ecb6