extract fact accumulator type
Summary:
The indexer adds facts to an accumulator structure which is defined in Symbol_add_fact along with the functions which use it.

This diff extracts this datatype to its own module Fact_acc with a restricted
interface. This enforces the datatype invariant and improves separation of concerns.

We move this module from `Symbol_add_facts` to `Symbol_predicate` where it more naturally belongs: in particular, we can restrict the interface of `Symbol_predicates` and encapsulate the invariant that `Fact_acc.t` is initialized with the list of all predicate.

Reviewed By: simonmar

Differential Revision: D34365298

fbshipit-source-id: 2fa7cfb200bc89eab2f09e4ae75d66872ad5f6f6