Improve subtyping for intersection on the left with arraykey
Summary:
There is a typechecking rule `t1 & t2 <: t3` iff `t1 <: t3 | not t2`. Previously, this only applied where `t2` was either `null` or `nonnull`. This adds `arraykey` and `not arraykey` to the possibilities, which will support a forthcoming change that adds more intersections with `arraykey`.

On occasion, we need to apply the rule above in reverse because the union on the right will lead us to take an incomplete step checking that either `t1 <: t3` or `t1 <: not t2`. But we do not want to risk getting stuck in a loop. Instead we only apply the rule in reverse where `t3` is a type variable. This is similar to how `t1 & nonnull <: t2` and `t1 <: ?t2` are handled.

One added test case fails if we do the first simplification but not the second.

By making more `not arraykey` types, this also surfaced a bug in negations. Negations (and disjointness) weren't taking into account that `string <: Stringish`, and were assuming that classes and primitives were disjoint. We fix that here too. The other test checks for that.

Reviewed By: CatherineGasnier

Differential Revision: D34273004

fbshipit-source-id: afad8730b533902f5be82aafb2e44ca0bac4601d