fix distinction between an abstract class and abstract classish
Summary:
Both classes and enum classes can be abstract, so we must take care to
use the right function to check if a classish is abstract (i.e. we don't
care about the kind of classish, just if that's abstract or not) and when
to check for an abstract class, explicitely.

Reviewed By: andrewjkennedy

Differential Revision: D34386201

fbshipit-source-id: acd2630337220fc4707508189e7143d99e3fa016