Ignore supportdyn in is/as constructs
Summary:
Currently we reject `is` or `as` with a `supportdyn` type. In this diff we make two changes

(1) We drop `supportdyn` when checking `is` and `as` constraints for well-formedness.
(2) We drop `supportdyn` from the expanded type when doing the refinement.

We don't support testing for support dynamic at runtime, but choose to support the use of `supportdyn<_>` in the types (or aliases) used in `is` and `as` expressions, dropping the support-dynamic from the refinement.

Differential Revision: D34386209

fbshipit-source-id: b7504806bc0d8dc00b117923c6691fccec8385a1