Add kind to FoldedClass
Summary: Indicates whether the FoldedClass represents a concrete class, an abstract class, an enum class (concrete or abstract), an interface, a trait, or a Hack enum (i.e., C-like enum). Directly copied over from the shallow representation (because users of a FoldedDeclProvider do not necessarily have access to a ShallowDeclProvider).

Reviewed By: edwinsmith

Differential Revision: D34396854

fbshipit-source-id: 52ac32127d783caf158f91800a66db0d92ee5326