Allow readonly closures defined to be mutable at definition time.
Summary:
When a user writes:
```
$f = readonly (int $x) ==> $x + 1;
```
today, they receive a *readonly reference* to a *readonly function*. I.e., a `readonly (readonly function(int):int)`. But when doing so, you don't really need to create a readonly reference to one: creating a mutable reference to a readonly function is fine.

This restriction was there initially because it was simpler to reason about, but I think the restriction does not need to be there.

This diff removes that restriction, which allows you to pass $f to functions that take regular functions.
Note that passing a mutable function to a readonly argument will still fail during execution of the closure for the runtime, and directly on the call in the typechecker (since function typehints are not enforced by the runtime).

This diff only changes it in the runtime to make it more loose, once HHVM is released we'll send it in the typechecker.

Reviewed By: oulgen

Differential Revision: D33573872

fbshipit-source-id: 88cea46e070cf8582c80733274ae6c71e34ea287