Add get_{fun,const,type} methods to FoldedDeclProvider
Summary:
Currently, decls of toplevel functions, global constants, and type aliases are inaccessible from typing logic. They are stored in a ShallowDeclProvider's backing cache, and the ShallowDeclProvider interface provides a means of reading them, but no such accessors exist in FoldedDeclProvider and TypingDeclProvider. We could give typing logic access to a ShallowDeclProvider, but this would give typing logic the ability to invoke `ShallowDeclProvider::get_class`, which is undesirable: an efficient implementation of `get_class` would require us to maintain a cache of shallow classes, which shouldn't be necessary in an environment where all folded classes are computed ahead of time (like distributed Zoncolan). In those environments, we'd like to save memory by discarding the shallow-class-cache after computing folded classes (because a folded class contains a superset of the information available in the corresponding shallow class).

Instead, communicate that the shallow-class-cache is unnecessary (but the FunDecl, ConstDecl, and TypedefDecl caches *are* necessary) by adding accessor functions to the FoldedDeclProvider trait (which simply delegate to the ShallowDeclProvider accessors added in the previous diff).

Reviewed By: shayne-fletcher

Differential Revision: D34406672

fbshipit-source-id: ea36215caaa8999fd45caf2e63b369c126719d74