Add --naming-table support to decl_file
Summary: Add a `NamingProvider` abstraction, to be used by `LazyShallowDeclProvider` only (folding and typechecking logic do not need access to a `NamingProvider`). Add an impl of `NamingProvider` for our existing `Names` type (which wraps a `rusqlite::Connection` and is used in the hh_server rearchitecture's `hh_decl` binary). Add an optional `--naming-table` argument to `decl_file` which points to a SQLite naming table (in the same format as the naming table database in hh_server's SQLite saved states). When a naming table is provided, use `LazyShallowDeclProvider` to decl-parse referenced parent classes on-the-fly. When no naming table is provided, use `EagerShallowDeclProvider` (and assume that all referenced names are in the set of files passed on the command line).

Reviewed By: shayne-fletcher

Differential Revision: D34402402

fbshipit-source-id: 42426f5c42e8d418aa581828cbb3338c1155673a