Error if meth_caller is given argument with '$'
Summary:
User request:

> I was messing around with meth_caller trying to pass it the mangled class name of a closure and noticed some confusion. Could you make HackC refuse to compile if the class name has a $ in it? Right now we just emit whatever you write into the name of the special function but HHVM always interprets the first $ in the name after the MethCaller$ prefix to be the end of the class name. Which means if you do meth_caller('C$x', 'memoize_impl') it's as if you wrote meth_caller('C', 'x$memoize_impl')

```
function test(): void {
  meth_caller('X$X', "foo");
}
```

While we already throw naming and type errors for these strings, we don't prevent the runtime from running the problematic code. This diff makes such a construction a parse error, stopping the runtime from running the code.

Note that this diff is not trying to ban strings that flow this way via variables as we cannot detect that statically. Instead, this type of construction throws a different naming error:

```
function test(): void {
  meth_caller($x, 'foo');
}

Naming[2026] The two arguments to meth_caller() must be:
 - first: ClassOrInterface::class
 - second: a single-quoted string literal containing the name of a non-static method of that class [1]
```

Reviewed By: periodic1236

Differential Revision: D34353927

fbshipit-source-id: 0863609723604bffb9624f96af2283f6867b4a7f