Use Vec<Instruct> in InstrSeq
Summary:
Every InstrSeq is constructed using a vec![...] which is immediately copied into
the arena and then destroyed. Now that InstrSeq is not part of the hhbc_ast FFI,
we can avoid all that copying by using the Vec directly.

This diff stops using arena allocation for InstrSeq, and unlocks a ton of `alloc`
plumbing for the next diff to clean up (and speed up a bit).

While here, rename several InstrSeq methods to idiomatic Rust names.

Reviewed By: aorenste

Differential Revision: D34363971

fbshipit-source-id: 3f4c950a0790ab1c41e07e8245e8457c221130fb