Flatten InstrSeq in Emitter instead of BytecodePrinter (3/3)
Summary:
Currently, all clients of HackCUnit must recursively traverse an InstrSeq holding bytecode instructions.
this is easy for Rust clients and a PITA for C++ clients. Instead, do it once in the emitter and have
HhasBody point to a single flat slice of instructions which is easy for both clients.

This removes the last InstrSeq in HackCUnit. This diff is suboptimal by itself
because it arena-allocates another copy of every instruction at the end of compilation.

However now that InstrSeq is no longer exposed to FFI, it can use Vec<Instruct>
internally and avoid lots of unnecessary vec->arena copying during compilation (D34363971).

Reviewed By: aorenste

Differential Revision: D34221812

fbshipit-source-id: 503b636e08d5321d9687a21a4cbd1a833ac48d26