Introduce ToOxidized trait for conversion from rupro types to oxidized_by_ref types
Summary:
We want to generate oxidized_by_ref values for the rupro folded decls in order to compare them against existing folded decls as measure of correctness. This diff introduces and implements the following trait for the relevant types (with some `unimplemented` to be filled out in later diffs).
```
pub trait ToOxidized<'a> {
    type Output: TrivialDrop + Clone + ToOcamlRep + 'a;

    fn to_oxidized(&self, arena: &'a bumpalo::Bump) -> Self::Output;
}
```

Differential Revision: D34325703

fbshipit-source-id: db2636ee7712bc0ebf4375a6be85a8a1b27c0cf1