ListIters for InstrSeq
Summary:
InstrIter is complex because it combines the functionality of walking an InstrSeq
tree iteratively with a separate stack, and flattening a leaf List to a stream of instructions.
We can separate those two features.

Lets have a family of iterators that only do the tree walking, each producing a stream
of *lists* instead of individual instructions.
* The new iterators are simpler (and faster)
* When we want to iterate instructions, we can use off-the-shelf flatten()
* By implementing iter(), iter_mut(), and IntoIter, we have a non-recursive iterator
for each type of ownership.

While here, a few more InstrSeq helpers became trivial enough to inline at callsites.

Reviewed By: aorenste, shayne-fletcher

Differential Revision: D34428738

fbshipit-source-id: 3015710e94bdba455fa0cfd1a616ae3b829c6f45