Add positions of shape field names
Summary:
These positions were omitted when we added the TshapeFieldName type. The comment explaining why was slightly wrong (see the edits to it in this diff): hash tables require Hash and Eq to agree, not Hash and Ord. However, Ord and PartialOrd are expected to agree with Eq and PartialEq, so it might still be best to avoid putting positions in this type (and writing an Ord impl which ignores them).

We do actually need these positions in order to convert decls back to the OCaml representation, so add them as part of the shape map's value, ShapeFieldType.

Differential Revision: D34456605

fbshipit-source-id: 92745f10c240fe0ef394cc54d62b4a1b4bc490e2