Make decl_provider optional, always use it if provided.
Summary:
This diff removes an unnecessary degree of freedom in the DeclProvider API.
Previously, an ENABLE_DECLS flag turned on functionality in hackc to request decls from
the decl-provider plug-in API. The provider could be real, or a stub that always returns NotFound.

Eventually we expect hackc to treat NotFound from a valid provider differently than the fall-back
behavior of no provider, so now only pass Some(provider) when it's a "real" decl provider,
and replace uses of the NoDeclProvider stub with None.

The FFI interface now uses `nullptr` to indicate None, and clients were refactored to avoid
constructing a provider at all, when not necessary.

Delete NoDeclProvider.

Reviewed By: shayne-fletcher

Differential Revision: D34306350

fbshipit-source-id: a5a61353d80a343d531dad4b4a72a8003cd0f630