prepare to move abstract enum class out of preview mode
Summary:
Starting to move AEC outside of unstable features.
This diff removes the actual check, and leaves the attribute in place.

This should avoid any error while this lands and gets into the HHVM release, and then wait for their release to be fully deployed. Once that's the case, I'll remove the attribute and notify people they can start stripping it from source code

Reviewed By: francesco-zappa-nardelli

Differential Revision: D34445687

fbshipit-source-id: 87aed2f6fdfff0ad3a744cc8a587e8bf97820c02