use iteration to build 86cinit()
Summary:
One of the recent bugs discovered with unnamed local variable numbering
had to do with the generation of 86cinit(). The bug is now fixed, but the codegen
logic was difficult to understand.

As a result of trying to understand it, this diff converts it from a recursive (but not
quite tail-recursive) algorithm to an iterative algorithm and removes an unnecessary
InstrSeq::clone() along the way.

Reviewed By: aorenste

Differential Revision: D34390330

fbshipit-source-id: 99fc49330c579c35ca7b741c4fc4b7bf7e1cb0fc