Property support
Summary:
A first diff implementing part of the HhasProgram parser for parsing properties.

The `86coeffects` prop we ignore here for now, because the logic that adds that in the runtime relies on methods, which are not yet supported.

Escaping logic follows very similar algorithm to current one in as.cpp

[RFC] Better ways to test this?
One way to test this is to compare the disassembly for the unit emitters generated both ways. To do this incrementally, it seems logical to only disassemble the parts of the unit emitter that are generated, and add fields as they are supported. Right now, it dumps properties and parts of classes.

Will flip flag back before land

Reviewed By: edwinsmith

Differential Revision: D31817885

fbshipit-source-id: fda86e1ea74d526c19bec161e73d49bbb77e2c01