Eliminate `ignore_errors` callback
Summary:
The `ignore_errors` callback was, pre-defunctionalization, a function which simply ignored the secondary errors it was passed. In the defunctionalized world, `Ignore` became a constructor on `Typing_error.Reasons_callback.t` and `Apply_reasons` a constructor on `Typing_error.t`.

This was fine since we only cared that *evaluating* `Apply_reasons(Ignore,...)` resulted in no reported errors but we would really like to be able to determine that there will be no errors without evaluating the `Typing_error.t`.

To that end, this diff removes the `Ignore` callback and makes the relevant functions accept a `Typing_error.Reasons_callback.t option` (instead of just `Typing_error.Reasons_callback.t`). Passing `None` now means that errors should be ignored.

Reviewed By: CatherineGasnier

Differential Revision: D34389757

fbshipit-source-id: c1fa866c150a0561f90c43dfef756311a58b50c4