Add ability to skip module initialization and systemlib parsing at startup
Summary:
Starting up a HHVM process can be pretty slow. Almost all of the time
is spent in two places: Initializing extensions, and compiling
systemlib. This is unnecessary for certain applications and can save a
lot of startup time. Add options to disable the initialization of
extensions or the compiling of systemlib during startup.

This will be useful when we migrate parsing to use subprocesses, as
the subprocesses don't need either of the above and it greatly speeds
up their initialization.

Reviewed By: mofarrell

Differential Revision: D34516390

fbshipit-source-id: 302ee629d27ec8e99bded002c06d61eef5bc3ce9